Introduce subprocess isolation: each sample runs hub + agent in separate venvs
Every sample now has two sub-projects:
- Orchestrator (<sample>/) depends only on xr-ai-launcher (stdlib-only)
- Worker (<sample>/worker/) depends only on xr-ai-agent + task libs

launcher/ gains three new primitives (all stdlib-only):
- ProjectLauncher: runs any uv project as a managed subprocess via
  `uv run --project`, yielding the asyncio.Process handle
- Process dataclass + StackLauncher + run_stack: declare the process
  sequence in orchestrator code; launcher auto-discovers each command's
  YAML config (<command>.yaml in the sample root) and passes --config

HubLauncher now delegates to ProjectLauncher, running the hub in
server-runtime's isolated venv regardless of the caller's venv.

All processes start concurrently — no ordering is required. ZMQ
reconnects automatically so ProcessorEndpoint tolerates the hub not
being ready at startup. run_stack is fail-fast: any process exit
terminates the whole stack.
