+++ agent-samples/vlm-agent/pyproject.toml
@@ -7,21 +7,11 @@ name = "vlm-agent"
 version = "0.1.0"
 requires-python = ">=3.11"
 dependencies = [
-    "xr-ai-agent",
     "xr-ai-launcher",
-    "numpy>=1.24",
-    "Pillow>=10.0",
-    "torch>=2.2",
-    "torchvision>=0.17",
-    "transformers>=4.49",
-    "accelerate>=0.30",
-    "qwen-vl-utils>=0.0.8",
-    "hf-transfer>=0.1.4",
 ]
 
 [tool.uv.sources]
-xr-ai-agent    = { path = "../../agent-sdk", editable = true }
-xr-ai-launcher = { path = "../../launcher",  editable = true }
+xr-ai-launcher = { path = "../../launcher", editable = true }
 
 [project.scripts]
 vlm_agent = "vlm_agent.__main__:run"
